requirements.txt
psutil>=5.9.0
schedule>=1.2.0
plyer>=2.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
pathlib>=1.0.1
pywin32>=306
anytree>=2.